Extract `NopComponent`.
