+++ src/webapp-builtins/package.json
@@ -8,7 +8,8 @@
   "exports": "./index.js",
   "imports": {
     "#x/*": "./export/*.js",
-    "#p/*": "./private/*.js"
+    "#p/*": "./private/*.js",
+    "#test/*": "./tests/*.js"
   },
 
   "dependencies": {
